Moved factory
- We have move the factory for resource type items
